refine model code
