+++ paddlemix/models/audioldm2/clap_module/htsat_model.py
@@ -14,60 +14,14 @@
 
 import paddle
 import paddle.nn as nn
-from itertools import repeat
-import collections.abc
+
 import math
 import warnings
 import random
 
 from .utils import do_mixup, interpolate, Spectrogram, LogmelFilterBank, SpecAugmentation
-
+from ..utils import to_2tuple, DropPath, Mlp
 from .feature_fusion import iAFF, AFF, DAF
-
-def _ntuple(n):
-    def parse(x):
-        if isinstance(x, collections.abc.Iterable):
-            return x
-        return tuple(repeat(x, n))
-
-    return parse
-
-to_1tuple = _ntuple(1)
-to_2tuple = _ntuple(2)
-to_3tuple = _ntuple(3)
-to_4tuple = _ntuple(4)
-to_ntuple = _ntuple
-
-def drop_path(x, drop_prob: float = 0.0, training: bool = False):
-    """Drop paths (Stochastic Depth) per sample (when applied in main path of residual blocks).
-    This is the same as the DropConnect impl I created for EfficientNet, etc networks, however,
-    the original name is misleading as 'Drop Connect' is a different form of dropout in a separate paper...
-    See discussion: https://github.com/tensorflow/tpu/issues/494#issuecomment-532968956 ... I've opted for
-    changing the layer and argument names to 'drop path' rather than mix DropConnect as a layer name and use
-    'survival rate' as the argument.
-    """
-    if drop_prob == 0.0 or not training:
-        return x
-    keep_prob = 1 - drop_prob
-    shape = (x.shape[0],) + (1,) * (
-        x.ndim - 1
-    )  # work with diff dim tensors, not just 2D ConvNets
-    random_tensor = keep_prob + paddle.rand(shape, dtype=x.dtype)
-    random_tensor.floor_()  # binarize
-    output = x.divide(keep_prob) * random_tensor
-    return output
-
-
-class DropPath(nn.Layer):
-    """Drop paths (Stochastic Depth) per sample  (when applied in main path of residual blocks)."""
-
-    def __init__(self, drop_prob=None):
-        super(DropPath, self).__init__()
-        self.drop_prob = drop_prob
-
-    def forward(self, x):
-        return drop_path(x, self.drop_prob, self.training)
-
 
 class PatchEmbed(nn.Layer):
     """2D Image to Patch Embedding"""
@@ -193,35 +147,6 @@ class PatchEmbed(nn.Layer):
             x = x.flatten(2).transpose([0, 2, 1])  # BCHW -> BNC
         x = self.norm(x)
         return x
-
-
-class Mlp(nn.Layer):
-    """MLP as used in Vision Transformer, MLP-Mixer and related networks"""
-
-    def __init__(
-        self,
-        in_features,
-        hidden_features=None,
-        out_features=None,
-        act_layer=nn.GELU,
-        drop=0.0,
-    ):
-        super().__init__()
-        out_features = out_features or in_features
-        hidden_features = hidden_features or in_features
-        self.fc1 = nn.Linear(in_features, hidden_features)
-        self.act = act_layer()
-        self.fc2 = nn.Linear(hidden_features, out_features)
-        self.drop = nn.Dropout(drop)
-
-    def forward(self, x):
-        x = self.fc1(x)
-        x = self.act(x)
-        x = self.drop(x)
-        x = self.fc2(x)
-        x = self.drop(x)
-        return x
-
 
 def _no_grad_trunc_normal_(tensor, mean, std, a, b):
     def norm_cdf(x):
@@ -1156,45 +1081,15 @@ class HTSAT_Swin_Transformer(nn.Layer):
 def create_htsat_model(audio_cfg, enable_fusion=False, fusion_type="None"):
     try:
         assert audio_cfg.model_name in [
-            "tiny",
             "base",
-            "large",
         ], "model name for HTS-AT is wrong!"
-        if audio_cfg.model_name == "tiny":
-            model = HTSAT_Swin_Transformer(
-                spec_size=256,
-                patch_size=4,
-                patch_stride=(4, 4),
-                num_classes=audio_cfg.class_num,
-                embed_dim=96,
-                depths=[2, 2, 6, 2],
-                num_heads=[4, 8, 16, 32],
-                window_size=8,
-                config=audio_cfg,
-                enable_fusion=enable_fusion,
-                fusion_type=fusion_type,
-            )
-        elif audio_cfg.model_name == "base":
+        if audio_cfg.model_name == "base":
             model = HTSAT_Swin_Transformer(
                 spec_size=256,
                 patch_size=4,
                 patch_stride=(4, 4),
                 num_classes=audio_cfg.class_num,
                 embed_dim=128,
-                depths=[2, 2, 12, 2],
-                num_heads=[4, 8, 16, 32],
-                window_size=8,
-                config=audio_cfg,
-                enable_fusion=enable_fusion,
-                fusion_type=fusion_type,
-            )
-        elif audio_cfg.model_name == "large":
-            model = HTSAT_Swin_Transformer(
-                spec_size=256,
-                patch_size=4,
-                patch_stride=(4, 4),
-                num_classes=audio_cfg.class_num,
-                embed_dim=256,
                 depths=[2, 2, 12, 2],
                 num_heads=[4, 8, 16, 32],
                 window_size=8,
